+++ Tests/JSONUtilities/InlineDecodingTests.swift
@@ -70,6 +70,71 @@ class InlineDecodingTests: XCTestCase {
     expectDecodeTypeArray(expectedBool)
   }
   
+  func testIncorrectEnum() {
+    
+    let dictionary = ["enum": "three"]
+    
+    do {
+      let _:MockParent.MockEnum = try dictionary.jsonKey("enumIncorrect")
+      XCTAssertThrowsError("Did not catch MandatoryKeyNotFound error")
+    }
+    catch let error {
+      let expectedError = DecodingError.MandatoryKeyNotFound(key: "enumIncorrect")
+      let actualError = error as! DecodingError
+      XCTAssert(expectedError == actualError)
+    }
+    
+    do {
+      let _:MockParent.MockEnum = try dictionary.jsonKey("enum")
+      XCTAssertThrowsError("Did not catch MandatoryRawRepresentableHasIncorrectValue error")
+    }
+    catch let error {
+      let expectedError = DecodingError.MandatoryRawRepresentableHasIncorrectValue(rawRepresentable: MockParent.MockEnum.self, rawValue: "three")
+      let actualError = error as! DecodingError
+      XCTAssert(expectedError == actualError)
+    }
+  }
+  
+  func test_decodingMandatoryEnumArray_withKey() {
+    let dictionary: JSONDictionary = ["enums": ["one", "!@1", "two"]]
+    
+    let decodedEnums: [MockParent.MockEnum] = try! dictionary.jsonKey("enums")
+    
+    let expectedEnums: [MockParent.MockEnum] = [.one, .two]
+    XCTAssertEqual(decodedEnums, expectedEnums)
+  }
+
+  
+  func test_decodingOptionalEnumArray_withKey() {
+    let dictionary: JSONDictionary = ["enums": ["one", "!@1", "two"]]
+    
+    let decodedEnums: [MockParent.MockEnum]? = dictionary.jsonKey("enums")
+    
+    let expectedEnums: [MockParent.MockEnum] = [.one, .two]
+    XCTAssertEqual(decodedEnums!, expectedEnums)
+  }
+  
+  func test_decodingMandatoryEnumArray_withoutKey() {
+    let dictionary: JSONDictionary = ["enums": ["one", "!@1", "two"]]
+    
+    do {
+      let _: [MockParent.MockEnum] = try dictionary.jsonKey("invalid_key")
+      XCTFail("Error not thrown")
+    } catch {
+      let expectedError = DecodingError.MandatoryKeyNotFound(key: "invalid_key")
+      XCTAssert(error as! DecodingError == expectedError)
+    }
+
+  }
+  
+  func test_decodingOptionalEnumArray_withoutKey() {
+    let dictionary: JSONDictionary = ["enums": ["one", "!@1", "two"]]
+    
+    let decodedEnums: [MockParent.MockEnum]? = dictionary.jsonKey("invalid_key")
+    
+    XCTAssertNil(decodedEnums)
+  }
+  
   func testDecodingOfJSONDictionaryArray() {
     
     let expectedValue: [JSONDictionary] = [["key1": "value1"], ["key2": "value2"]]